// frontend/src/generated/endpoints/collections/collections.ts
export type createCollectionResponse201 = {
  data: EnvelopeCollectionSummary
  status: 201
}

export type createCollectionResponseDefault = {
  data: ErrorModel
  status: Exclude<HTTPStatusCodes, 201>
}

export type createCollectionResponseSuccess = (createCollectionResponse201) & {
  headers: Headers;
};
export type createCollectionResponseError = (createCollectionResponseDefault) & {
  headers: Headers;
};

export type createCollectionResponse = (createCollectionResponseSuccess | createCollectionResponseError)
// ...
    export type deleteCollectionResponse204 = {
  data: void
  status: 204
}

export type deleteCollectionResponseDefault = {
  data: ErrorModel
  status: Exclude<HTTPStatusCodes, 204>
}

export type deleteCollectionResponseSuccess = (deleteCollectionResponse204) & {
  headers: Headers;
};
export type deleteCollectionResponseError = (deleteCollectionResponseDefault) & {
  headers: Headers;
};

export type deleteCollectionResponse = (deleteCollectionResponseSuccess | deleteCollectionResponseError)
// ...
export type updateCollectionResponse200 = {
  data: EnvelopeCollectionSummary
  status: 200
}

export type updateCollectionResponseDefault = {
  data: ErrorModel
  status: Exclude<HTTPStatusCodes, 200>
}

export type updateCollectionResponseSuccess = (updateCollectionResponse200) & {
  headers: Headers;
};
export type updateCollectionResponseError = (updateCollectionResponseDefault) & {
  headers: Headers;
};

export type updateCollectionResponse = (updateCollectionResponseSuccess | updateCollectionResponseError)